Windows version compiling and removed type warnings int/ssize_t.
Datagram sockets building on Windows

Added sock_address::family()

+++ examples/tcp6/tcp6echo.cpp
@@ -69,7 +69,7 @@ int main(int argc, char* argv[])
 		}
 
 		sret.resize(s.length());
-		int n = conn.read_n(&sret[0], s.length());
+		ssize_t n = conn.read_n(&sret[0], s.length());
 
 		if (n != (int) s.length()) {
 			cerr << "Error reading from TCP stream: "
